// client/src/dicts/roles-and-permissions.ts
//ROLES
const globalAdmin = 'GLOBAL_ADMIN'
const headOfDepartment = 'HEAD_OF_DEPARTMENT'
const hr = 'HR'
const user = 'USER'

export enum PERMISSION_TYPES {
  manageCreatePositionView = 'MANAGE_CREATE_POSITION_VIEW',
  manageVacancyFormView = 'MANAGE_VACANCY_FORM_VIEW',
  managePositionsView = 'MANAGE_POSITIONS_VIEW',
  manageProposalsView = 'MANAGE_PROPOSALS_VIEW',
  manageVacanciesView = 'MANAGE_VACANCIES_VIEW',
  manageCandidatesView = 'MANAGE_CANDIDATES_VIEW',
  manageReserveView = 'MANAGE_RESERVE_VIEW',
  manageInternsView = 'MANAGE_INTERNS_VIEW',
  manageInterviewsView = 'MANAGE_INTERVIEWS_VIEW',
  manageSendMessageView = 'MANAGE_SEND_MESSAGE_VIEW',
  manageStatisticsView = 'MANAGE_STATISTICS_VIEW',
  manageInterviewFormView = 'MANAGE_INTERVIEW_FORM_VIEW',
  manageCandidatesProfileView = 'MANAGE_CANDIDATES_PROFILE_VIEW',

}

export enum ROLES_AND_PERMISSIONS {

    GLOBAL_ADMIN= 'GLOBAL_ADMIN',
    HEAD_OF_DEPARTMENT= 'HEAD_OF_DEPARTMENT',
    HR='HR',
    USER= 'USER'

}